+++ CHANGELOG.md
@@ -23,13 +23,17 @@ to [Semantic Versioning]. Full commit history is available in the
 
 #### Removed
 
-### 1.2.2 (2024-XX-XX)
+### 1.2.2 (2024-12-31)
 
 #### Added
 
 - Add MuData Minification option to {class}`~scvi.model.TOTALVI` {pr}`3061`.
+- Add Support for MPS usage in mac {pr}`3100`.
+- Add support for torch.compile before train (EXPERIMENTAL) {pr}`2931`.
 - Add support for Numpy 2.0 {pr}`2842`.
 - Changed scvi-hub ModelCard and add criticism metrics to the card {pr}`3078`.
+- MuData support for {class}`~scvi.model.MULTIVI` via the method
+    {meth}`~scvi.model.MULTIVI.setup_mudata` {pr}`3038`.
 
 #### Fixed
 
@@ -53,8 +57,6 @@ to [Semantic Versioning]. Full commit history is available in the
     validation set, if available. {pr}`3036`.
 - Add `batch_key` and `labels_key` to `scvi.external.SCAR.setup_anndata`. {pr}`3045`.
 - Implemented variance of ZINB distribution. {pr}`3044`.
-- MuData support for {class}`~scvi.model.MULTIVI` via the method
-    {meth}`~scvi.model.MULTIVI.setup_mudata` {pr}`3038`.
 - Support for minified mode while retaining counts to skip the encoder.
 - New Trainingplan argument `update_only_decoder` to use stored latent codes and skip training of
     the encoder.